asm: Interface update - Added entry and label structs
The assembler will be a linear two-pass system. The first pass will decode
instructions and data declarations, and also keep track of labels; the
second pass will resolve all references. Finally, theentries will be written
out to an asm_binary struct that can be immediately executed or written to a
file.

Note: the asm_binary struct is not (currently) binary-equivalent to the
executable format. A specific procedure must be used to transform between
them.